requirements.txt
absl-py==1.0.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.28
dataclasses==0.8
decorator==5.1.1
docker-pycreds==0.4.0
Flask==2.0.1
gitdb==4.0.9
GitPython==3.1.18
glfw==2.5.1
google-auth==2.6.4
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gtimer==1.0.0b5
gym==0.21.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipdb==0.13.9
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mujoco-py==1.50.1.68
numpy==1.19.5
oauthlib==3.2.0
pandas==1.1.5
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.1.0
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==1.5.7
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smmap==5.0.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
torch==1.10.0+cu113
torchvision==0.11.0+cu113
tqdm==4.63.0
traitlets==4.3.3
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.9
# Editable install with no version control (viskit==0.1)
-e /home/zhwang/viskit-master
wandb==0.12.11
wcwidth==0.2.5
websocket-client==1.3.1
Werkzeug==2.0.3
wrapt==1.14.0
wurlitzer==3.0.2
yacs==0.1.8
yaspin==2.1.0
zipp==3.6.0
